Move `services` singletons into `src/services` folder
- create `src/services` folder
- deprecate notice for `core/data/expression/service`
- deprecate notice for `core/data/ows/service`
- deprecate notice for `core/data/proxy/service`
- deprecate notice for `core/data/query/service`
- deprecate notice for `core/data/search/rservice`
- deprecate notice for `core/data/routerservice`
- deprecate notice for `core/editing/changesmanager`
- deprecate notice for `core/history_router`
- deprecate notice for `core/iframe/routerservice`
- deprecate notice for `core/iframe/services/app/service`
- deprecate notice for `core/iframe/services/plugins/editing/service`
- deprecate notice for `core/relations/relationsservice`
- deprecate notice for `core/router`
- deprecate notice for `core/task/service`
- deprecate notice for `core/workflow/workflowstack`
- deprecate notice for `gui/gui`
- deprecate notice for `gui/querybuilder/service`
- consolidate singletons defintions within a single folder
